mv _download_and_parse into base warehouse
