+++ dephell/repositories/warehouse/_api.py
@@ -1,8 +1,6 @@
 # built-in
 import asyncio
 from logging import getLogger
-from pathlib import Path
-from tempfile import TemporaryDirectory
 from typing import Dict, Iterable, List, Optional, Tuple, Union
 from urllib.parse import urlparse
 from xmlrpc.client import ServerProxy
@@ -21,13 +19,7 @@ from ...config import config
 from ...exceptions import PackageNotFoundError, InvalidFieldsError
 from ...models.author import Author
 from ...models.release import Release
-from ..base import Interface
-
-
-try:
-    import aiofiles
-except ImportError:
-    aiofiles = None
+from ._base import BaseWarehouse
 
 
 logger = getLogger('dephell.repositories')
@@ -62,7 +54,7 @@ def _process_url(url: str) -> str:
 
 
 @attr.s()
-class WareHouseRepo(Interface):
+class WareHouseRepo(BaseWarehouse):
     name = attr.ib(default='pypi')
     url = attr.ib(type=str, factory=lambda: config['warehouse'], converter=_process_url)
     prereleases = attr.ib(type=bool, factory=lambda: config['prereleases'])  # allow prereleases
@@ -313,38 +305,3 @@ class WareHouseRepo(Interface):
                     except FileNotFoundError as e:
                         logger.warning(e.args[0])
         return ()
-
-    async def _download_and_parse(self, *, url: str, converter) -> Tuple[str, ...]:
-        with TemporaryDirectory() as tmp:
-            async with ClientSession() as session:
-                async with session.get(url) as response:
-                    if response.status != 200:
-                        raise ValueError('invalid response: {} {} ({})'.format(
-                            response.status, response.reason, url,
-                        ))
-                    path = Path(tmp) / url.rsplit('/', maxsplit=1)[-1]
-
-                    # download file
-                    if aiofiles is not None:
-                        async with aiofiles.open(str(path), mode='wb') as stream:
-                            while True:
-                                chunk = await response.content.read(1024)
-                                if not chunk:
-                                    break
-                                await stream.write(chunk)
-                    else:
-                        with path.open(mode='wb') as stream:
-                            while True:
-                                chunk = await response.content.read(1024)
-                                if not chunk:
-                                    break
-                                stream.write(chunk)
-
-            # load and make separated dep for every env
-            root = converter.load(path)
-            deps = []
-            for dep in root.dependencies:
-                for env in dep.envs.copy():
-                    dep.envs = {env}
-                    deps.append(str(dep))
-            return tuple(deps)
